Add serdes field_serializers (#13081)
## Summary & Motivation

Add `field_serializers` functionality to serdes.

This allows custom field serializer logic to be scoped to a single
field. The primary motivation for this is to allow serialization of a
`metadata` dict as a list of metadata_entries and vice versa-- see
upstack: https://github.com/dagster-io/dagster/pull/12770

## How I Tested These Changes

New unit tests.