[dagit] Refresh health in the asset sidebar if live data updates (#13201)
## Summary & Motivation

This PR resolves #13023. If you're viewing the asset graph and have an
asset selected, it's partition health bar is shown in the sidebar.
Previously, the health bar would not refresh when the asset's partition
counts updated on the graph because the health bar uses a separate
GraphQL query that is not refreshed periodically.

The `usePartitionHealthData` hook already accepts a "cache hint" that
allows us to use lightweight polling to trigger a heavy refresh of the
health bar (on the asset details page). This PR wires it up on the asset
graph as well, and moves some of the code to a shared helper.

## How I Tested These Changes

We don't have tests for the asset graph sidebar yet. To test this I
launched runs of an asset from the asset graph and observed that the
counts on the graph and the counts in the sidebar are now in sync.

I also changed the partition definition in code and verified that both
the asset graph and the asset sidebar updated to reflect the new number
of partitions in the asset. (The case in #12422)

Co-authored-by: bengotow <bgotow@elementl.com>