[5/n][resources ui] Expose resource-asset dependencies in repo snapshot (#13007)
## Summary

Walks assets defined in a repo and builds a list of the resources they
depend on, binding this info to the resources so we can surface it in
the UI in the 'uses' page:

```python
class MyResource(ConfigurableResource):
    a_str: str

@asset
def my_asset(foo: MyResource):
    pass

@asset
def my_other_asset(foo: MyResource):
    pass

defs = Definitions(
    assets=[my_asset, my_other_asset],
    resources={"foo": MyResource(a_str="foo")}
)
```

Here, the `asset_keys_using` attribute of `foo` will have the asset keys
for `my_asset`, `my_other_asset`.

See the stacked PR for the UI impl.

## Test Plan

Unit test.