[refactor] Update internal representation of metadata to always use metadata dict (#12770)
### Summary & Motivation

Internal companion PR: https://github.com/dagster-io/internal/pull/5257

Update internal representation of `metadata` from `metadata_entries` to
the dict form we have been pushing to users for almost a year.

- Remove `metadata_entries` from all non-public constructor args
- Introduce `MetadataSerializer` custom field serializer to serialize
metadata dict to list of metadata entries and deserialize list of
metadata entries to metadata dict (for backcompat). Apply to all serdes
classes holding `metadata_entries`

**Context**

More than a year ago, it was decided that Dagster's preferred
representation of metadata was to be the metadata dict, as opposed to
the metadata entries list:

```
### metadata dict (yes)
{
    "foo": MetadataValue.text("bar"),
    "baz": MetadataValue.int(1)
}

### metadata entries list (no)
[
    MetadataEntry("foo", value=MetadataValue.text("bar")),
    MetadataEntry("baz", value=MetadataValue.int(1)),
]
```

Reasons for this transition:

- metadata dict is more user-friendly
- we deprecated the `description` field on `MetadataEntry` so
`MetadataEntry` no longer had any greater expressive power over a dict

There was a hitch in this plan though-- `metadata_entries:
Sequence[MetadataEntry]` was widely used as an internal representation
of metadata throughout the codebase, including on many
serdes-whitelisted objects. So we couldn't just replace the metadata
entry list everywhere. Instead, we decided on a temporary solution of:

- offer `metadata` kwarg on user-facing APIs to allow them to pass a
metadata dict
- internally convert to list of metadata entries
- keep using list of metadata entries everywhere for internal repr

This has a few problems:

- we have a lot of constructors that take either `metadata` or
`metadata_entries`. Sometimes `metadata_entries` is marked private
(`_metadata_entries`) other times not. It's confusing.
- we inconsistently reexpose metadata as a dict-- but if we do this, and
we are using a `MetadataEntry` list for the internal representation,
then we have to reconvert to a dict

All in all the current situation is inconsistent and introduces
unnecessary complexity to the code.

This PR should be viewed as a stepping stone to the complete elimination
of `MetadataEntry` and passing metadata as `List[MetadataEntry]`, which
is done upstack: https://github.com/dagster-io/dagster/pull/12769

### How I Tested These Changes

Existing test suite.