Guard against upstream croniter bug causing scheduler ticks to get stuck in a time loop
Summary:
croniter seems to have a lot of trouble with spring dst transitions - attempted to dig in there and fix them but the problems were too numerous to really make progress. This PR is a tactical retreat that at least eliminates the worst problem, which is a tick looping over and over again at (this year) 1AM on 3/27 for a 2AM schedule. The test here documents the new behavior, which is still not great but at least advances and reaches an OK state again.

Test Plan: BK, test that was failing before