[Checkbox] stopPropagation for `onClick` (#13282)
## Summary & Motivation

The Checkbox component wraps the checkbox input and its label in a
`<label>` element. This means that when the checkbox is clicked, any
parent `onClick` handlers will fire twice (see this codesandbox for a
demonstration:
https://codesandbox.io/s/muddy-https-6zypxg?file=/src/index.js) .

To fix this I move the `onClick` handler onto the `label` element, and I
call `stopPropagation` when clicking on the checkbox input directly. see
for demonstration of the fix:
https://codesandbox.io/s/nice-star-ppf0hk?file=/src/index.js