[Dynamic partition requests 2/3] SensorResult (#12962)
This PR introduces a `SensorResult` object that wraps all of the return
objects from the sensor. This lays the groundwork for allowing users to
return a dynamic partition request object that mutates a dynamic
partitions definition as part of a `SensorResult`.

The `SensorResult` object allows for users to pass in:
- a cursor for the next time the sensor is evaluated
- and either a skip reason or a combination of run requests and pipeline
run reactions. At least one of these objects must be provided.

However, the cursor update is only applied for regular sensors defined
via `@sensor`. The other types of sensors (multi asset sensors, asset
sensors, run status sensor) manage their cursors with dagster, so if a
cursor update is attempted via a `SensorResult` an error is thrown.