Fix min on empty list in data time resolution (#13237)
## Summary & Motivation

Unclear why, but in some cases, this list is empty. Guard against this
scenario

Edit: this was happening because of a weirdly-structured query in
get_event_log_records, which would do the wrong thing if rows were
deleted. The root issue has been corrected, but this doesn't hurt as a
safeguard

## How I Tested These Changes