[ui] Upgrade to Blueprint 4 (#13308)
## Summary & Motivation

Upgrade to Blueprint 4. This seems like a fairly straightforward change,
with the `bp3` -> `bp4` change being the main difference.

https://github.com/palantir/blueprint/wiki/Blueprint-4.0

I'm looking at upgrading us to React 18, and this is likely to be a bit
simpler with updated Blueprint types.

There will probably be a long tail of minor visual or behavioral changes
that we can track down, but this seems pretty solid.

## How I Tested These Changes

TS, lint, jest.

Load the app and navigate around. Verify that components look and behave
correctly.

Load storybook, verify that components look correct.