test(dbt): use `dbt-duckdb` adapter (#13198)
## Summary & Motivation
Use duckdb for all of local testing. This allow us the added benefit of
actually being able to take advantage of our out-of-the-box io manager
for testing `dagster-dbt` specific features, as well removes our
dependency on Docker/Postgres.

The test suite still takes a while to run though, so more investigation
needs to be done to test changes fast. But this is a step in that
direction.

## How I Tested These Changes
pytest