[test-api-update] rename utility solids to op equivalents (#13172)
## Summary & Motivation

This is a migration of a large number of tests to job/op/graph APIs. It
started with an attempt to remove `execute_solid` and friends, but that
ended up to be pulling on a string that had to be further unraveled to
get to a passing state-- so this now includes not only `execute_solid`
conversions but also a full conversion of the test-project repo and
updating of associated tests.

## How I Tested These Changes

Test suite, much of which is changed.