AutoMaterializePolicy namedtuple (#13362)
I'm breaking up Owen's draft PR
https://github.com/dagster-io/dagster/pull/13205/files#diff-912be2c7b2dfba3ccde0782f7db04b79eaeb8a52f80a6e2eb3a62c2aa9ddde09
into smaller pieces. I want to land the API without the changes to the
sensor behavior. It's not clear to me that we need to change how the
sensor works- it may be easier to just have this api work with the
daemon and leave the sensor untouched.