Better handle ecsrunlauncher with use_current_ecs_task_config=False but no task_definition set
Summary:
We were implicitly assuming that if you set use_current_ecs_task_config to False, you must want to configure the task definition in some way, but that is not always the case. Make it possible to launch a run using a 'default' task definition that pulls nothing from the run.

We actually had a test for this case, but it stil had access to the env var that is set when you are inside the ECS task, so the test passed incorrectly.

Test Plan: BK test that would fail before