Use specific events to trigger a "Preparing" state instead of assuming that any logged event means that the step has started (#13182)
Summary:
This is intended to address a problem where the Dagit UI incorrectly
thinks that a step is being prepared after its upstream event fails so
we log a "Dependencies for step XXX failed. Not executing." events.

Since
https://github.com/dagster-io/dagster/commit/e163de1e7d1927199242b87f25b1839b818908d7
we have added additional structured events that explicitly indicate that
a step has started. Use those to determine that the step is being
prepared - until then, leave the state null.

Furthermore, source the set of steps that haven't executed from the
graph, rather than purely the event log - since a step that didn't
execute may not have logged an event at all.

Test Plan:
Launch this run:
```
import time

from dagster import asset

@asset
def a():
    time.sleep(45)

@asset
def c(b):
    pass

@asset
def b():
    raise Exception("Oops")

```

Before this change, c is shown in a Preparing / Not executed state in
the right hand bar. After this change, it is not shown in the right hand
bar at all.

Launch a normal run where every step succeeds or fails, no changes

Are there tests related to dynamic orchestration that I should check,
can we count on the nodes in the graph matching up with the keys in the
metadata?

## Summary & Motivation

## How I Tested These Changes

Co-authored-by: OwenKephart <owen@elementl.com>