Allow jobs with unpartitioned assets to be partitioned (#13472)
Source assets in the job are not returned as part of the asset
selection, which causes the removed error to be raised.

For now, this PR removes the error to unblock users. In the future, we
can update the error messaging to handle when partitioned source assets
are also selected.