[ui] Virtualize the Run timeline (#13483)
## Summary & Motivation

Virtualize the Run timeline view. This should significantly improve the
experience for users with huge numbers of jobs and/or runs.

## How I Tested These Changes

Load app with thousands of jobs and thousands of runs. Verify that the
run timeline loads a whole lot quicker than before, and that virtualized
scrolling and rendering works as expected. Collapse and expand code
location headers, verify that behavior is correct.

Verify that the empty state and loading states look correct, and that
everything still behaves correctly in cases where there are far fewer
jobs and runs.