[dagit] Split Stale tag into several more specific tags, update catalog status column (#13462)
## Summary & Motivation

This PR is linked to the discussion here:
https://github.com/dagster-io/dagster/discussions/13102,
https://github.com/dagster-io/internal/discussions/5381


- Replaces the word "Late" with "Overdue"

- Replaces generalized uses of the concept "is stale" with "has
unpropagated changes"

- Splits the single "Stale" tag into three. They can appear above the
asset node if they refer to it's inputs, or inside the asset node if the
referenced assetkey is the current one.

- The asset catalog now has a "Status" column instead of "Last
Materialized" which matches the asset graph. To keep things sane and
always consistent, the catalog and assetNode share a bit of the
rendering code.



![image](https://user-images.githubusercontent.com/1037212/230829759-d5022ee3-f8ae-4318-a4e9-099d9bfac894.png)


![image](https://user-images.githubusercontent.com/1037212/230977415-32ccccab-a7c4-4b69-acd5-b5e7bfae44c9.png)


## Testing Approach

I used our storybooks to do most of the required AssetNode changes and
update the asset catalog table. Because these are just new rendering
states for largely existing states, i think this is a good way to
verifying correctness.

I verified with Sean that the upstream tags are working as expected on a
test DAG.

---------

Co-authored-by: bengotow <bgotow@elementl.com>