Add resource support to observable source assets (#11993)
### Summary & Motivation

Allows provision of resources to observable source assets via either
`required_resource_keys` (old-style) or Pythonic resource args.

Threading `required_resource_keys` through both
`@observable_source_asset` and `SourceAssetDefinition` was awkward. If
we allow config schema for observation functions (which we should) it
will only get worse. The solution used here is to make
`_required_resource_keys` a private arg on `SourceAsset` to avoid
committing on some API questions discussed below.

Another possible approach would be to allow source asset to take a
`node_def` (matching the `AssetsDefinition` API) instead of an
`observe_fn`. That would technically be a break we wouldn't normally
make, since `observe_fn` is not marked as an experimental arg on
`SourceAsset`-- but I'd argue it is widely understood to be experimental
since `@observable_source_asset` is, which is probably what anyone using
this feature is using anyway.

### How I Tested These Changes

New unit tests