[dagit] Ungate resources UI (#13517)
## Summary

Ungates the Resources UI in Dagit. Leaves the feature flag and turns it
into a toggle for showing/hiding resources in the sidebar - now that
they're visible in the overview page and deployments page I think this
is less necessary and could create clutter. Leaving it as an experiment
feels pretty noncommittal, but happy to remove entirely.

## Test Plan

Tested locally.