[ui] Add asset groups to global search (#13515)
## Summary & Motivation

Resolves #13231.

Add asset groups to global search, and repair the "many locations" check
that determines whether to show the code location name with the result.

## How I Tested These Changes

Load app, open search. Type the name of an asset group, verify that it
appears. Navigate to it, verify success. Verify that the code location
names also appear correctly even though there are many repositories in a
single location -- they are treated like separate code locations.