Allow @job resources_def argument to take any object, not just a ResourceDefinition (#13556)
## Summary & Motivation

Now that Definitions resources automatically bind resources to jobs, it makes sense to have the behavior between them consistent, including automatic coercion. Without this change one cannot pass the same dictionary of objects to `Definitions` and `job`, which doesn't make too much sense.

This is also important for setting up a change needed to ConfigurableResource and friends in order to make them not inherit from ResourceDefinition.

## How I Tested These Changes

BK and included test.