Make rules in MakeConfigCacheable much more strict (#13581)
## Summary & Motivation

MakeConfigCacheable is a class that allows us attach variables directly
to Pydantic classes that are otherwise immutable. Before we had a fairly
blanket policy of allowing any variable starting with "_" to be set on
these classes. That was very broad and it would be easier for our users
to get themselves into trouble.

Here we make this much more strict and only fields ending in
`__internal__` can be set on these classes. For the classes the depend
on this behavior I extracted out a separate class so that only these
variables exist per class.

## How I Tested These Changes

BK