[ui] Remove custom ID function from Apollo cache setup (#13594)
## Summary & Motivation

Cleaning up something we should have taken care of a while ago. Instead
of defining custom `id` behavior for cached singletons when we define
the app cache, just add some `id` fields to the actual GraphQL types.

Also removed a custom branch for `RegularType` and `CompositeType`,
neither of which are typenames that seem to exist now.

## How I Tested These Changes

Buildkite. Load app, click around and verify that everything looks fine,
including in the cache display in the Apollo dev tools.