[ui] Fix half-clipped headers on repo-scoped asset table (#13656)
## Summary & Motivation

This PR renames the asset catalog table stories and adds a story for the
repo-scoped asset table, which has slightly different columns and
folding behavior. It fixes an issue with the repo-scoped asset table
defining one too many columns and adjusts the column widths slightly.

Thankfully these use the same virtualized rows and the storybooks are
almost the same.

Note: I also added an "id" value to some Workspace mocks, because the
storybook was throwing exceptions without an id. I'm not 100% sure all
of the places I added it are required, but I think specifying a constant
value is correct and we shouldn't let our old faker-based Apollo mocks
randomize the value.

## How I Tested These Changes


![image](https://user-images.githubusercontent.com/1037212/232147007-8fac39c3-6e67-4665-850a-f36109e5fec1.png)

![image](https://user-images.githubusercontent.com/1037212/232147025-0507aa79-83ee-4bbd-9dc0-edc68904a31c.png)

---------

Co-authored-by: bengotow <bgotow@elementl.com>