remove usage of stale in some docs (#13658)
## Summary & Motivation

Updates docs to respect the revised staleness ontology, which does not
use the word "Stale".

As part of this change, I removed the tab on the asset details page that
deals with "Upstream changed". This was already out of date.

This PR also includes a fix to the imports in one of the
AssetObservation examples. I can pull that out into a separate change if
helpful.

## How I Tested These Changes