remove MetadataEntry APIs (#12769)
### Summary & Motivation

Delete all APIs accepting `MetadataEntry`, but keep `MetadataEntry`
itself. Add a formal deprecation to `MetadataEntry` for removal in 2.0.

The reason for this somewhat awkward state is that we deprecated all
APIs where you can _use_ `MetadataEntry` a year ago, but forgot to
deprecate `MetadataEntry` itself.

The deprecate usage sites for `MetadataEntry` that are being removed in
this PR are:
- passing `metadata_entries` to user-constructable events
(`AssetObservation`, `AssetMaterialization`, `ExpectationResult`,
`TypeCheck`, `Failure`, `Output`, `DynamicOutput`)
- passing `metadata_entries` to `DagsterType`
- returning a value from `event_metadata_fn` in `dagster-pandas`

In all these cases, passing a list of `MetadataEntry` has been emitting
deprecation warnings for approximately a year-- this functionality was
supposed to be removed in 0.15.0/0.16.0, but we didn't get around to it.

Note that this PR was enabled by
https://github.com/dagster-io/dagster/pull/12770, which updated our
internal representation to never use `MetadataEntry`.

### How I Tested These Changes

Existing BK suite w/ some removed tests, plus a new test for a
deprecation warning on `MetadataEntry` construction.