Rename DataDogClientResource to DatadogResource; DataDogClient to DatadogClient (#13700)
Per our new naming conventions, rename DataDogClientResource to
DatadogResource. And DataDogClient to DatadogClient as that appears to
be more consistent with Datadog's capitalization norms.

Despite this url's (https://github.com/DataDog/datadogpy ) indication to
the contrary, it refers to the "Datadog Python Library" in the body of
its readme, so we are following that.

BK