redirect scheduling assets guide to auto-materializing concept page (#13709)
## Summary & Motivation

This is a great guide and I feel sad to get rid of it, but it's now
redundant with other content that we've introduced on declarative
scheduling, and it's obsolete given the recent changes we've introduced
around staleness and AutoMaterializePolicys.

This PR removes it and adds redirects. In the future, I'd like to
plunder some of the content and use it to beef up the auto-materializing
guide.

## How I Tested These Changes

(cherry picked from commit aa1b6aa1d9db1e94af45f5053ab79287fde0eee7)