add test for hourly-to-daily backfill (#13698)
## Summary & Motivation

I wrote this test to try to replicate some behavior I observed while
testing backfills manually. It ended up passing - i.e. not replicating
the behavior I had observed, but it still seemed worthwhile to include.

The backfill test harness didn't have the ability to set a
`current_time` or target a particular partition subset, which was needed
for this test, so I added it.

## How I Tested These Changes