[ui] Cache code location reload errors after mutation (#13699)
## Summary & Motivation

When reloading a code location, cache the mutation response. This should
update the UI to show the correct error.

The goal here is to resolve an issue where a code location has multiple
errors surfaced when loading it. After fixing one error and reloading,
the fixed error doesn't go away. This is because the error is tracked
via the top-level workspace query, but not updated by the mutation. By
writing to the cache, the error dialog updates with the new (as yet
unfixed) error right away.


https://user-images.githubusercontent.com/2823852/232617176-c9515797-7871-49cf-abbd-791e6555525b.mov

## How I Tested These Changes

Raise exceptions from two `@repository` objects in `toys/repo.py`, load
UI. Navigate to code locations, open error dialog. Comment out one
exception, click "Reload". Verify that the next exception shows up in
the dialog. Comment it out, click Reload. Verify that the code location
loads successfully.