[ui] Fix quick-loading code location toast (#13724)
## Summary & Motivation

Fix an edge case in monitoring code location status. When a code
location is ready so quickly that we never notice it loading, we show a
toast indicating that it has been added. However, we were showing the
`id` of the status entry object, not the name of the location itself.
The result was a weird toast that looked broken:


![image](https://user-images.githubusercontent.com/2823852/232839266-5263ed1c-1e90-4751-8a19-91a42a0de99e.png)


Repair this by getting the location name from the entry map. This
codepath is pretty rare, which I guess is why we haven't noticed it
before.

<img width="347" alt="Screenshot 2023-04-18 at 11 13 02 AM"
src="https://user-images.githubusercontent.com/2823852/232839152-d12f53ef-1925-4142-81cd-38e2c816d827.png">

Also fix the toaster button for the `primary` toast, which is showing up
as blue on a gray background since the recent Blueprint v4 upgrade.

## How I Tested These Changes

Force this codepath when loading a new code location to ensure that the
toast shows up with the correct location name.

View toaster in Storybook to sanity check color change.