Cache the returned ResourceDefinition from ConfigurableResource instances (#13807)
## Summary


https://elementl-workspace.slack.com/archives/C03D094F23G/p1682010604398169

> ```dagster._core.errors.DagsterInvalidDefinitionError: Conflicting
versions of resource with key 'fivetran' were provided to different
assets. When constructing a job, all resource definitions provided to
assets must match by reference equality for a given key.```

The `ResourceDefinitions` here are produced by the
`ConfigurableResource`, this PR caches the generated
`ResourceDefinition` (which was itself broken out in the late change in
#13567)

## Test Plan

Added new unit test w/ multiple connections that replicates the error,
see that change fixes.