[2/n][Runs page] Update filtering to use new component (#13758)
## Summary & Motivation

We're updating the Runs page to the new filtering component.

There is still a quirk due to the `Loading` component recreating its
children from scratch in a way that they lose their internal state. This
causes the dropdown to close whenever the page refreshes. I need to dig
into this further but it doesn't look like a quick fix for now.

There are also further visual changes to make but I will do those in a
follow-up.

## How I Tested These Changes

Jest  + manual testing

https://www.loom.com/share/7bf0cfd841284b02a27e73511b95b92d