Add partition_key field to RunStatusSensorContext (#13811)
## Summary & Motivation
Follow up from discussion here:
https://github.com/dagster-io/dagster/discussions/13283

This is a very minor change to expose the partition as a top level field
in `RunStatusSensorContext`. Previously this information was only
available by checking the tags that are available on the `DagsterRun`
object within the `RunStatusSensorContext`.

## How I Tested These Changes
Updated unit tests with a scenario that covers the case when a run
status sensor is defined on a partitioned job.