Disable run_request_for_partition with dynamic partitions (#13817)
Each call of `run_request_for_partition` for a dynamic partition
requires querying the instance for the existent dynamic partitions,
which is expensive and can overwhelm the db.

This PR disables this behavior by removing the instance arg and throwing
an error. I wish initially I'd added an `experimental` tag to the
instance arg, but I think only dynamic partitions users should be
passing in an instance and thus will encounter this breaking change.

If we don't feel comfortable directly removing the instance arg, I'm
happy to add the arg back and just not pass the instance through to the
`get_partition` method.