[ui] Performance improvements for global search (#13833)
## Summary & Motivation

Introduce a couple of performance improvements on global search.

- Debounce queries while typing.
- This delays actually executing the search so that we avoid expensive
searches that we don't need.
- Move Fuse.js searching into Web Workers.
- For very large data collections, Fuse.js is pretty slow. By pushing
the cost to a worker thread, we avoid letting Fuse block the main
thread.
- Use two of these Web Workers for global search, one for the "primary"
query (most of the defined objects: jobs, schedules, etc.) and one for
the "secondary" query (currently assets).
- Since most definitions should be available soon after the page loads,
the primary query should (hopefully) be ready soon after search is
opened.
- With separate threads for these data collections, we can start showing
primary results even before the secondary GraphQL query is complete, or
before the secondary Fuse query is complete.

Each of these improvements helps a bit on their own, but they work best
in conjunction, so I'm adding both in this PR.

When the user opens search:

- Two lazy GraphQL queries are started, primary and secondary.
- When a GraphQL query completes, create a search worker and populate it
with the results. The GraphQL query results may be cached initially, in
which case the worker should already have a collection to search on.
- While either GraphQL query is loading, a loading spinner appears in
the search input.
- When the user types, perform a debounced search. Results arrive async
from the worker. If results arrive that don't match the user's current
querystring, discard them.

I also made some tweaks to the Fuse options to improve surfacing exact
matches.

Since we have other searchable pages throughout the app, we can use
search workers elsewhere as well and get Fuse fuzzy searching in a bunch
of places.

## How I Tested These Changes

Created Storybook stories to simulate:

- Basic search with not-too-many objects
- Fast primary query, slow secondary query
- Normal primary query, seconcary query with tens of thousands of
objects

In each case, open search and verify (visually and with Performance
profiling in dev tools) that querying and rendering are snappy.

Repeat search in app to sanity check.