Push dynamic partition insert down to Postgres
Our base sql event log implementation inserts dynamic partitions by
first finding existing ones and then inserting only new ones.

Generally, this is fine. But if you try to create a lot of dynamic
partitions in a single call and a lot of the dynamic partitions already
exist, there's a lot of overhead in pulling such a large list into
memory.

SQLAlchemy doesn't have a generalized abstraction for insert if not
exist but it does have dialect specific ones for certain dbs. This
overloads `add_dynamic_partitions` for Postgres to push the behavior
down to the db layer.